backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
playwright>=1.48.0
supabase>=2.11.0
pydantic>=2.10.0
python-dotenv>=1.0.0
google-api-python-client>=2.158.0
openai>=1.59.0
httpx>=0.27.0
APScheduler>=3.10.4
cryptography>=43.0.0
beautifulsoup4>=4.12.0
pytz>=2024.1
selenium>=4.27.0
webdriver-manager>=4.0.0

